+++ packages/base/.storybook/main.ts
@@ -16,7 +16,6 @@ const config: StorybookConfig = {
     getAbsolutePath("@storybook/addon-essentials"),
     getAbsolutePath("@storybook/addon-interactions"),
     getAbsolutePath("@storybook/addon-a11y"),
-    "storybook-addon-performance"
   ],
   framework: {
     name: getAbsolutePath("@storybook/react-vite"),