+++ packages/base/src/components/Buttons/DropdownButton/index.ts
@@ -1,0 +1,4 @@
+export { default as DropdownButton } from './DropdownButton';
+export * from './DropdownButton';
+export * from './DropdownButton.types';
+export * from './useDropdownButton';