+++ packages/base/src/components/Buttons/SplitButton/index.ts
@@ -1,0 +1,4 @@
+export { default as SplitButton } from './SplitButton';
+export * from './SplitButton';
+export * from './SplitButton.types';
+export * from './useSplitButton';