+++ packages/base/src/components/Buttons/index.ts
@@ -1,3 +1,5 @@
 // SHOPLFLOW_DOCS_IGNORE
 export * from './Button';
 export * from './IconButton';
+export * from './DropdownButton';
+export * from './SplitButton';