+++ packages/base/src/components/Comboboxs/NumberCombobox/index.ts
@@ -1,0 +1,3 @@
+export { default as NumberCombobox } from './NumberCombobox';
+export * from './NumberCombobox';
+export * from './NumberCombobox.types';