+++ packages/base/src/components/Datepickers/AnnualDatepicker/AnnualDatepicker.stories.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 
 import type { StoryFn } from '@storybook/react';
 import { Stack } from '../../Stack';