+++ packages/base/src/components/Inputs/Input/Input.types.ts
@@ -1,6 +1,14 @@
 import type { InputHTMLAttributes } from 'react';
-import type { DisableProps, RightElementProps, ErrorProps } from '../../../utils/type/ComponentProps';
+import type { DisableProps, RightElementProps, ErrorProps, SizeVariantProps } from '../../../utils/type/ComponentProps';
 import type { BorderRadiusTokens } from '../../../styles';
+import type { $Values } from '@shoplflow/utils';
+
+export const InputSizeVariants = {
+  S: 'S',
+  M: 'M',
+} as const;
+
+export type InputSizeVariantType = $Values<typeof InputSizeVariants>;
 
 export interface InputProps
   extends InputOptionProps,
@@ -8,10 +16,17 @@ export interface InputProps
     DisableProps,
     RightElementProps {}
 
-export interface InputOptionProps extends ErrorProps {
+export interface InputOptionProps extends ErrorProps, SizeVariantProps<InputSizeVariantType> {
   width?: string;
   maxLength?: number;
+  /**
+   * 최소 넓이를 지정
+   * @description 64px;
+   */
+  minWidth?: string;
   borderRadius?: BorderRadiusTokens;
   customNumberInputHeight?: string;
   onClear?: () => void;
+  gap?: CSSStyleDeclaration['gap'];
+  initIsFocused?: boolean;
 }