+++ packages/base/src/components/index.ts
@@ -24,3 +24,4 @@ export * from './Tabs';
 export * from './Pagination';
 export * from './ToggleButton';
 export * from './Datepickers';
+export * from './Comboboxs';